unify droid adapters
